feat: move banner type to type dir and move banner data to src/data
